Add count for Audiobooks Chapter to restart download

+++ main.py
@@ -16,14 +16,11 @@ def createFolder(directory):
 
 def getAllAudiobooks():
     audiobooks = []
-
     url = "https://iceportal.de/api1/rs/page/hoerbuecher"
     response = requests.get(url, headers=cfg.headers)
-
     # extract titles
     json_data = json.loads(response.text)
     items = json_data["teaserGroups"][0]["items"]
-
     for item in items:
         name = str(item["navigation"]["href"])
         itemtype = str(item["subtitle"])
@@ -31,20 +28,20 @@ def getAllAudiobooks():
         # check if itemtype is not Podcast
         if itemtype != "Podcast":
             audiobooks.append(name)
-
     return audiobooks
-
 
 def downloadAudiobook(title):
     titleshort = title.split("/")[2]
     print("Downloading audiobook: {}".format(title))
     boolTest = False
     jsonFilePath = "audiobooks/{}/{}.json".format(titleshort, titleshort)
+    countFilePath = "audiobooks/{}/{}.count".format(titleshort, titleshort)
 
     if os.path.exists(jsonFilePath):
         print("audiobook exists")
         return
 
+ 
     url = "https://iceportal.de/api1/rs/page{}".format(title)
     responseChapter = requests.get(url, headers=cfg.headers)
 
@@ -52,9 +49,6 @@ def downloadAudiobook(title):
 
     # extract chapters
     json_data = json.loads(responseChapter.text)
-
-    with open(jsonFilePath, "w") as jsonFile:
-        jsonFile.write(json.dumps(json_data, indent=4))
     
     playlist = json_data["files"]
 
@@ -72,37 +66,50 @@ def downloadAudiobook(title):
     
     for chapter in playlist:
         chapterPath = chapter["path"]
-
-        url = "https://iceportal.de/api1/rs/audiobooks/path{}".format(
-            chapterPath)
-        responseDownloadPath = requests.get(
-            url, headers=cfg.headers, cookies=cfg.cookies)
-
+        url = "https://iceportal.de/api1/rs/audiobooks/path{}".format(chapterPath)
+        responseDownloadPath = requests.get( url, headers=cfg.headers, cookies=cfg.cookies)
         path = json.loads(responseDownloadPath.text)["path"]
         downloadPath.append(path)
 
     # download each track
+    count = 0
+    if os.path.exists(countFilePath):
+        with open(countFilePath, "r") as countFile:
+            count = int(countFile.read())
+            if count >= len(downloadPath):
+                print("audiobook exists")
+                return
+
     for counter, track in enumerate(downloadPath):
         print("{}/{}".format(counter+1, len(downloadPath)))
-
         url = "https://iceportal.de{}".format(track)
         ext = track.split(".")[-1]
         audio = requests.get(url)
-
         savePath = "audiobooks/{}/{}_".format(titleshort, titleshort)+str(counter+1)+"."+ext
+        
+        if count == (counter+1) and os.path.exists(savePath):
+            os.remove(savePath)
+
+        if os.path.exists(savePath):
+            continue
+
         with open(savePath, "wb+") as code:
             code.write(audio.content)
 
+        with open(countFilePath, "w") as countFile:
+            countFile.write(str(counter+1))
+
+    with open(jsonFilePath, "w") as jsonFile:
+        jsonFile.write(json.dumps(json_data, indent=4))
+
+
 def getAllPDFs():
     PDFs = []
-
     url = "https://iceportal.de/api1/rs/page/zeitungskiosk"
     response = requests.get(url, headers=cfg.headers, cookies=cfg.cookies)
-
     # extract titles
     json_data = json.loads(response.text)
     items = json_data["teaserGroups"][0]["items"]
-
     for item in items:
         name = str(item["navigation"]["href"])
         # check if itemtype is not Podcast
@@ -115,19 +122,13 @@ def downloadPDF(title):
     print("Downloading PDF: {}".format(title))
     url = "https://iceportal.de/api1/rs/page{}".format(title)
     responseChapter = requests.get(url, headers=cfg.headers, cookies=cfg.cookies)
-
-
     # extract chapters
     json_data = json.loads(responseChapter.text)
-
     itemurl = str(json_data["navigation"]["href"])
     itemdate = str(json_data["date"])
     titleshort = str(json_data["segment"])
-
     createFolder('./zeitungskiosk/{}'.format(titleshort))
-   
     url = "https://iceportal.de/{}".format(itemurl)
-
     savePath = "zeitungskiosk/{}/{}-{}".format(titleshort,itemdate,titleshort)+".pdf"
     fileDonePath = "zeitungskiosk/{}/{}-{}".format(titleshort,itemdate,titleshort)+".done"
     if os.path.exists(fileDonePath):
@@ -140,7 +141,6 @@ def downloadPDF(title):
     # create done file
     with open(fileDonePath, "w") as code:
             code.write("done")
-
 
 # MAIN
 # extract all audiobooks
